test(dedupe): consolidate table_formatting tests (5 → 2 files)
Cluster c1e4c1ed0ffb: merge similar_names test files for table_formatting

Files consolidated:
- test_table_formatting.py (NEW): cells + clipboard + truncate_id (142 lines)
- test_table_formatting_data.py (NEW): sorting + timestamp (227 lines)

Removed:
- test_table_formatting_cells.py
- test_table_formatting_clipboard.py
- test_table_formatting_sorting.py
- test_table_formatting_timestamp.py
- test_table_formatting_truncate_id.py

All 50 tests pass.
