+++ tests/unit/gpt_trader/tui/utilities/test_table_formatting.py
@@ -1,10 +1,18 @@
-"""Tests for DataTable cell formatting utilities."""
+"""Tests for DataTable cell formatting, clipboard, and ID utilities."""
 
+import subprocess
 from decimal import Decimal
+from unittest.mock import MagicMock
 
+import pytest
 from rich.text import Text
 
-from gpt_trader.tui.utilities.table_formatting import create_numeric_cell, format_table_cell
+from gpt_trader.tui.utilities.table_formatting import (
+    copy_to_clipboard,
+    create_numeric_cell,
+    format_table_cell,
+    truncate_id,
+)
 
 
 class TestCreateNumericCell:
@@ -78,3 +86,57 @@ class TestFormatTableCell:
         """Format None value."""
         result = format_table_cell(None)
         assert result == "--"  # Implementation returns "--" for None
+
+
+class TestCopyToClipboard:
+    """Tests for copy_to_clipboard function."""
+
+    def test_copy_on_macos(self, monkeypatch: pytest.MonkeyPatch) -> None:
+        """Test clipboard copy on macOS."""
+        mock_run = MagicMock()
+        mock_run.return_value.returncode = 0
+        monkeypatch.setattr(subprocess, "run", mock_run)
+        # This test is platform-dependent, so we just verify no crash
+        result = copy_to_clipboard("test text")
+        # Result depends on platform and available tools
+        assert isinstance(result, bool)
+
+    def test_copy_empty_string(self) -> None:
+        """Copy empty string doesn't crash."""
+        result = copy_to_clipboard("")
+        assert isinstance(result, bool)
+
+
+class TestTruncateId:
+    """Tests for truncate_id function."""
+
+    def test_long_id_truncated(self) -> None:
+        """Long ID is truncated to last N characters."""
+        result = truncate_id("abc123def456ghi789", length=8)
+        assert result == "56ghi789"
+
+    def test_short_id_unchanged(self) -> None:
+        """Short ID (shorter than length) is unchanged."""
+        result = truncate_id("short", length=8)
+        assert result == "short"
+
+    def test_exact_length_unchanged(self) -> None:
+        """ID exactly at length is unchanged."""
+        result = truncate_id("12345678", length=8)
+        assert result == "12345678"
+
+    def test_empty_string(self) -> None:
+        """Empty string returns empty string."""
+        result = truncate_id("")
+        assert result == ""
+
+    def test_none_returns_empty(self) -> None:
+        """None returns empty string."""
+        result = truncate_id(None)  # type: ignore[arg-type]
+        assert result == ""
+
+    def test_default_length(self) -> None:
+        """Default length is 8."""
+        result = truncate_id("0123456789abcdef")
+        assert result == "89abcdef"
+        assert len(result) == 8
